2025-04-cli-demo/pyproject.toml
[project]
name = "2025-04-cli-demo"
version = "0"
description = "A demo of using Logfire with an internal CLI tool"
requires-python = ">=3.13"
dependencies = [
    "opentelemetry-api>=1.20.0",
    "logfire>=3.14.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
extend-select = [
    "Q",
    "RUF100",
    "C90",
    "UP",
    "I",
]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
isort = { combine-as-imports = true }
mccabe = { max-complexity = 15 }

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# don't format python in docstrings, pytest-examples takes care of it
docstring-code-format = false
quote-style = "single"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportUnnecessaryTypeIgnoreComment = true
include = [
    "**/*.py",
]
venvPath = ".venv"

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
